feat: move components to lib and fix imports
