+++ lib/components/Pagination.tsx
@@ -1,11 +1,11 @@
-import { tw, tx } from '@helpwave/common/twind/index'
-import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
-import { useTranslation } from '@helpwave/common/hooks/useTranslation'
-import type { Languages } from '@helpwave/common/hooks/useLanguage'
-import ChevronLast from '@helpwave/common/icons/ChevronLast'
-import ChevronLeft from '@helpwave/common/icons/ChevronLeft'
-import ChevronFirst from '@helpwave/common/icons/ChevronFirst'
-import ChevronRight from '@helpwave/common/icons/ChevronRight'
+import { tw, tx } from '../twind'
+import type { PropsWithLanguage } from '../hooks/useTranslation'
+import { useTranslation } from '../hooks/useTranslation'
+import type { Languages } from '../hooks/useLanguage'
+import ChevronLast from '../icons/ChevronLast'
+import ChevronLeft from '../icons/ChevronLeft'
+import ChevronFirst from '../icons/ChevronFirst'
+import ChevronRight from '../icons/ChevronRight'
 
 type PaginationTranslation = {
   of: string
