+++ lib/components/examples/Title.tsx
@@ -1,8 +1,8 @@
 import type { FunctionComponent } from 'react'
-import { tw } from '@helpwave/common/twind/index'
-import { useTranslation } from '@helpwave/common/hooks/useTranslation'
-import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
-import type { Languages } from '@helpwave/common/hooks/useLanguage'
+import { tw } from '../../twind'
+import { useTranslation } from '../../hooks/useTranslation'
+import type { PropsWithLanguage } from '../../hooks/useTranslation'
+import type { Languages } from '../../hooks/useLanguage'
 
 type TitleTranslation = {
   welcome: string,
