+++ lib/components/user_input/Menu.tsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react'
 import type { PropsWithChildren, ReactNode, RefObject } from 'react'
-import { tw, tx } from '@helpwave/common/twind/index'
-import { useOutsideClick } from '@helpwave/common/hooks/useOutsideClick'
+import { tw, tx } from '../../twind'
+import { useOutsideClick } from '../../hooks/useOutsideClick'
 
 type MenuProps<T> = PropsWithChildren<{
   trigger: (onClick: () => void, ref: RefObject<T>) => ReactNode,
