+++ lib/icons/Helpwave.tsx
