+++ tasks/components/MobileInterceptor.tsx
@@ -2,7 +2,7 @@ import { tw } from '@helpwave/common/twind'
 import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
 import { useTranslation } from '@helpwave/common/hooks/useTranslation'
 import Link from 'next/link'
-import HelpwaveLogo from '../icons/HelpwaveRect'
+import HelpwaveLogo from '@helpwave/common/icons/HelpwaveRect'
 import type { NextPage } from 'next'
 import { AndroidView, IOSView } from 'react-device-detect'
 import { getConfig } from '../utils/config'
