+++ tasks/components/OrganizationForm.tsx
@@ -1,9 +1,9 @@
-import { tw, tx } from '@helpwave/common/twind/index'
+import { useState } from 'react'
+import { tw, tx } from '@helpwave/common/twind'
 import type { Languages } from '@helpwave/common/hooks/useLanguage'
 import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
 import { useTranslation } from '@helpwave/common/hooks/useTranslation'
-import { Input } from './user_input/Input'
-import { useState } from 'react'
+import { Input } from '@helpwave/common/components/user_input/Input'
 
 type OrganizationFormTranslation = {
   general: string,
