+++ tasks/components/SubtaskView.tsx
@@ -1,9 +1,9 @@
 import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
 import { useTranslation } from '@helpwave/common/hooks/useTranslation'
-import { tw } from '@helpwave/common/twind/index'
+import { tw } from '@helpwave/common/twind'
 import type { SubTaskDTO } from '../mutations/room_mutations'
 import SimpleBarReact from 'simplebar-react'
-import { Button } from './Button'
+import { Button } from '@helpwave/common/components/Button'
 import Add from '@helpwave/common/icons/Add'
 import { SubtaskTile } from './SubtaskTile'
