+++ tasks/components/WardForm.tsx
@@ -1,8 +1,8 @@
-import { tw, tx } from '@helpwave/common/twind/index'
+import { tw, tx } from '@helpwave/common/twind'
 import type { Languages } from '@helpwave/common/hooks/useLanguage'
 import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
 import { useTranslation } from '@helpwave/common/hooks/useTranslation'
-import { Input } from './user_input/Input'
+import { Input } from '@helpwave/common/components/user_input/Input'
 import { useState } from 'react'
 
 type WardFormTranslation = {
