+++ tasks/components/cards/OrganizationCard.tsx
@@ -2,8 +2,8 @@ import { tw } from '@helpwave/common/twind'
 import type { Languages } from '@helpwave/common/hooks/useLanguage'
 import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
 import { useTranslation } from '@helpwave/common/hooks/useTranslation'
-import type { CardProps } from './Card'
-import { Card } from './Card'
+import type { CardProps } from '@helpwave/common/components/Card'
+import { Card } from '@helpwave/common/components/Card'
 import Email from '@helpwave/common/icons/Email'
 import { AvatarGroup } from '../AvatarGroup'
 import { Edit } from 'lucide-react'
