+++ tasks/components/layout/TaskDetailView.tsx
@@ -1,12 +1,12 @@
 import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
 import type { TaskDTO } from '../../mutations/room_mutations'
 import { useTranslation } from '@helpwave/common/hooks/useTranslation'
-import { tw } from '@helpwave/common/twind/index'
-import { ToggleableInput } from '../user_input/ToggleableInput'
-import { Textarea } from '../user_input/Textarea'
-import { Select } from '../user_input/Select'
+import { tw } from '@helpwave/common/twind'
+import { ToggleableInput } from '@helpwave/common/components/user_input/ToggleableInput'
+import { Textarea } from '@helpwave/common/components/user_input/Textarea'
+import { Select } from '@helpwave/common/components/user_input/Select'
 import { TaskStatusSelect } from '../user_input/TaskStatusSelect'
-import { Button } from '../Button'
+import { Button } from '@helpwave/common/components/Button'
 import { SubtaskView } from '../SubtaskView'
 import { X } from 'lucide-react'
