+++ tasks/pages/login.tsx
@@ -4,15 +4,15 @@ import Head from 'next/head'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import Cookies from 'js-cookie'
-import { tw } from '@helpwave/common/twind/index'
+import { tw } from '@helpwave/common/twind'
 import { loginWithCredentials } from '../utils/login'
-import { Input } from '../components/user_input/Input'
-import { Checkbox } from '../components/user_input/Checkbox'
-import { Button } from '../components/Button'
+import { Input } from '@helpwave/common/components/user_input/Input'
+import { Checkbox } from '@helpwave/common/components/user_input/Checkbox'
+import { Button } from '@helpwave/common/components/Button'
 import type { Languages } from '@helpwave/common/hooks/useLanguage'
 import { useTranslation } from '@helpwave/common/hooks/useTranslation'
 import type { PropsWithLanguage } from '@helpwave/common/hooks/useTranslation'
-import HelpwaveLogo from '../icons/Helpwave'
+import HelpwaveLogo from '../../lib/icons/Helpwave'
 import titleWrapper from '../utils/titleWrapper'
 
 type LoginTranslation = {
