+++ tasks/pages/profile.tsx
@@ -2,11 +2,11 @@ import { useEffect, useState } from 'react'
 import type { PropsWithChildren } from 'react'
 import type { NextPage } from 'next'
 import Head from 'next/head'
-import { tw } from '@helpwave/common/twind/index'
+import { tw } from '@helpwave/common/twind'
 import { useAuth } from '../hooks/useAuth'
 import { useRouter } from 'next/router'
-import { Button } from '../components/Button'
-import { Input, noop } from '../components/user_input/Input'
+import { Button } from '@helpwave/common/components/Button'
+import { Input, noop } from '@helpwave/common/components/user_input/Input'
 import { Avatar } from '../components/Avatar'
 import { UserCard } from '../components/cards/UserCard'
 import { PageWithHeader } from '../components/layout/PageWithHeader'
